Java code for Honeycomb Word Search
The program takes two command line input arguments. The first is the path of a text file containing the number of layers in the honeycomb followed by carriage-return delimited sets of uppercase letters that make up each layer of cells (Honeycomb input file). The second is the path of a text file containing carriage-return delimited valid strings with 2 or more uppercase letters (Dictionary input file).

This program finds all valid strings (from the dictionary input file) that can be found by starting at an arbitrary cell and building a word by moving to adjacent connected cells one at a time and prints them all to stdout in alphabetical order. Cells cannot be used twice in the formation of a single string. If a valid string can be found via multiple paths, that string is printed just once.